Refactor Intl polyfills and remove from ES5 entrypoints
