+++ src/common/datetime/first_weekday.ts
@@ -1,11 +1,7 @@
 import { getWeekStartByLocale } from "weekstart";
 import { FrontendLocaleData, FirstWeekday } from "../../data/translation";
 
-import { polyfillsLoaded } from "../translations/localize";
-
-if (__BUILD__ === "latest" && polyfillsLoaded) {
-  await polyfillsLoaded;
-}
+import "../../resources/intl-polyfill";
 
 export const weekdays = [
   "sunday",
