+++ src/common/datetime/relative_time.ts
@@ -1,11 +1,7 @@
 import memoizeOne from "memoize-one";
 import { FrontendLocaleData } from "../../data/translation";
-import { polyfillsLoaded } from "../translations/localize";
+import "../../resources/intl-polyfill";
 import { selectUnit } from "../util/select-unit";
-
-if (__BUILD__ === "latest" && polyfillsLoaded) {
-  await polyfillsLoaded;
-}
 
 const formatRelTimeMem = memoizeOne(
   (locale: FrontendLocaleData) =>
